Move config to internal & Refactor env vars
